# Game-Hub/game-Hub-be/game-Hub/src/main/java/gamehub/game_Hub/Request/ReportRequest.java
package gamehub.game_Hub.Request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ReportRequest {

  private Long reason;
  private String message;


}