# Game-Hub/game-Hub-fe/src/app/services/models/user-public-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BadgeResponse } from '../models/badge-response';
import { CardColorResponse } from '../models/card-color-response';
import { GameResponseShort } from '../models/game-response-short';
import { GenreResponse } from '../models/genre-response';
import { LevelResponse } from '../models/level-response';
import { LocationResponse } from '../models/location-response';
export interface UserPublicResponse {
  badges?: Array<BadgeResponse>;
  bannerImage?: string;
  bannerType?: string;
  bio?: string;
  cardColor?: CardColorResponse;
  favoriteGames?: Array<GameResponseShort>;
  favoriteGenres?: Array<GenreResponse>;
  friendsCount?: number;
  joinedDate?: string;
  level?: LevelResponse;
  libraryCount?: number;
  location?: LocationResponse;
  playRecently?: Array<GameResponseShort>;
  predefinedBannerPath?: string;
  profileColor?: string;
  recommendedGames?: Array<GameResponseShort>;
  status?: 'ONLINE' | 'OFFLINE' | 'AWAY';
  userId?: number;
  userProfilePicture?: string;
  username?: string;
  wishlistCount?: number;
}