# Game-Hub/game-Hub-fe/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Marcel,",
      "url": "https://google.com",
      "email": "test@gmail.com"
    },
    "license": {
      "name": "License name",
      "url": "https://random-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local environment"
    },
    {
      "url": "https://api.example.com/v1",
      "description": "Production environment"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/report/user/{userId}": {
      "post": {
        "tags": [
          "report-controller"
        ],
        "operationId": "reportUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/profile/update-profile": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "updateUserProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/profile/update-bio": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "updateBio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/profile/predefined/banner": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "setPredefinedBanner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/profile/image": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "uploadProfileImage",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/profile/custom/banner": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "uploadBannerImage",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/profile/add/favorite/genres": {
      "post": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "updateFavoriteGenres",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "uniqueItems": true,
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/profile/user/{userId}": {
      "get": {
        "tags": [
          "user-profile-controller"
        ],
        "operationId": "getUserPublic",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReportRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "enum": [
              "SK",
              "CZ",
              "PL",
              "HU",
              "US",
              "UNKNOWN"
            ]
          },
          "cardColorId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BannerRequest": {
        "type": "object",
        "properties": {
          "bannerPath": {
            "type": "string"
          },
          "bannerType": {
            "type": "string"
          }
        }
      },
...
      "GenreResponse": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "LevelResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "type": "integer",
            "format": "int64"
          },
          "levelColor": {
            "type": "string"
          }
        }
      },
...
      "BadgeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "iconPath": {
            "type": "string"
          }
        }
      },
      "CardColorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "colorName": {
            "type": "string"
          },
          "colorCode": {
            "type": "string"
          }
        }
      },
      "GameResponseShort": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "gameCoverImage": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "LocationResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "iconPath": {
            "type": "string"
          }
        }
      },
      "UserPublicResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "AWAY"
            ]
          },
          "joinedDate": {
            "type": "string"
          },
          "friendsCount": {
            "type": "integer",
            "format": "int32"
          },
          "libraryCount": {
            "type": "integer",
            "format": "int32"
          },
          "wishlistCount": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "$ref": "#/components/schemas/LevelResponse"
          },
          "badges": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BadgeResponse"
            }
          },
          "playRecently": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResponseShort"
            }
          },
          "favoriteGames": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResponseShort"
            }
          },
          "favoriteGenres": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponse"
            }
          },
          "recommendedGames": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResponseShort"
            }
          },
          "userProfilePicture": {
            "type": "string",
            "format": "byte"
          },
          "bannerImage": {
            "type": "string",
            "format": "byte"
          },
          "bannerType": {
            "type": "string"
          },
          "predefinedBannerPath": {
            "type": "string"
          },
          "profileColor": {
            "type": "string"
          },
          "cardColor": {
            "$ref": "#/components/schemas/CardColorResponse"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}